switch to Qt-provided qmlformat (from jesperhh/qmlfmt)
qmlformat is now part of the official Qt distribution. However,
it seems to have some warts and seems to act like beta software.
we will trial it and consider opening some bug reports. Switching
back to jesperhh/qmlfmt is a very real possibility. We were very
happy with jesperhh/qmlfmt