qmake: enable more compiler warnings on mac/ios
this was not done previously because I had not figured out how to
tell the mac/ios compiler chain that Qt headers are 'system' headers.
finally got that figured out.

by making the warnings (which can fail the build in CI due to Werror)
more consistent across platforms, there will be less "churn" and wasted
time in CI, because it should be less common to upload something that
builds without errors on one platform, only to find out that the same
new code produces errors on the other platform.