enhance the visual QML debug helper: toggle via environment, not code
DebugRectangle was added in commit 1d945aea58a93cd2a7731a50d55c5c66f1a87e36
as an easy way to quickly add visual borders to QML elements, to aid in
tracking down problems with alignment/sizing/layout.

The current commit further improves the convenience of DebugRectangle.
Now you can leave the DebugRectangle instances embedded in the QML code,
and you don't need to edit 'visible' on anything or rebuild the QRC of
your app in order to toggle visibility of the debug borders.

Now, whether you launch your application or whether you only launch your
QML using qmlscene, the debug borders will show up whenever you:

    export QT_QUICK_CONTROLS_MATERIAL_ACCENT="#0B610B" # our 'magic' sentinel value

Ideally we would invent our own environment variable for this purpose.
However, since we want this to work when using qmlscene, we are reliant
100% on only what is achievable in pure QML/Javascript (no support from
C++ viewModel classes we write). In pure QML, there is no way to read
from an environment variable.  Therefore, we use this "clever hack"
which piggy-backs on an environment variable that is read by the C++
code of the QML internal engine itself. The QML internals (included in
both your app and qmlscene) read the value of QT_QUICK_CONTROLS_MATERIAL_ACCENT
and make it available in QML as `Material.accent`. In this way, we
can test the value of `Material.accent` in pure QML to know whether
the environment variable was set.