rename libapp. in Qt android, our app itself ('main') becomes libapp. hence:
commit c5d3e1a actually broke the android apk, but nothing in the
android compilation toolchain flagged it as any kind of build failure.
when you attempted to install the APK on a device and launch the app,
you would see errors related to symbols from src/lib_app, such as:

     AndroidRuntime: java.lang.UnsatisfiedLinkError: dlopen failed: cannot locate symbol "_ZTIN7project11LoggingTagsE" referenced by "/data/app/com.mycompany.myapp-3-x2YbBFVsbkrA4s4BbH9A==/lib/arm64/libapp_arm64-v8a.so"...

when you use qmake to produce an APK, what (in the desktop world)
is normally our executable named 'app' instead becomes a lib, which
android-qmake names libapp. but since (as of our commit c5d3e1a) we coincidentally
already had a helper library which we explicitly named libapp, this
'main' libapp for android was clobbering our helper lib, leading
to the UnsatisfiedLinkError(s) for symbols in our helper library.